Move cmd/pivit/generate.go -> pkg/pivit/generate.go
