+++ pkg/pivit/generate.go
@@ -1,4 +1,4 @@
-package main
+package pivit
 
 import (
 	"crypto"
@@ -14,15 +14,14 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/cashapp/pivit/pkg/pivit"
 	"github.com/cashapp/pivit/pkg/pivit/utils"
 	"github.com/cashapp/pivit/pkg/pivit/yubikey"
 	"github.com/go-piv/piv-go/piv"
 	"github.com/pkg/errors"
 )
 
-// commandGenerate generates a new key pair and certificate signing request
-func commandGenerate(slot string, isP256, selfSign, generateCsr, assumeYes bool, pinPolicy piv.PINPolicy, touchPolicy piv.TouchPolicy) error {
+// CommandGenerate generates a new key pair and certificate signing request
+func CommandGenerate(slot string, isP256, selfSign, generateCsr, assumeYes bool, pinPolicy piv.PINPolicy, touchPolicy piv.TouchPolicy) error {
 	yk, err := yubikey.Yubikey()
 	if err != nil {
 		return err
@@ -113,7 +112,7 @@ func commandGenerate(slot string, isP256, selfSign, generateCsr, assumeYes bool,
 			return errors.Wrap(err, "parse self-signed certificate")
 		}
 
-		if err := pivit.ImportCertificate(cert, yk, managementKey, slot); err != nil {
+		if err := ImportCertificate(cert, yk, managementKey, slot); err != nil {
 			return errors.Wrap(err, "import self-signed certificate")
 		}
 	} else if generateCsr {
